pyproject.toml
[tool.poetry]
name = "atividade ativa - camara dos deputados"
version = "1.0.0"
description = "Atividade ativa"
authors = ["Gilliard Fernandes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.42.0"
numpy = "^2.2.3"
pandas = "^2.2.3"
requests = "^2.32.3"
pysqlite3 = "^0.5.4"
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.2"
plotly-express = "^0.4.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
